refactor: move streaming utils
