+++ projects/console-web-front/src/components/streaming/DeviceToolbar.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image';
 import React from 'react';
 import { HomeOutlined, LockOutlined, MenuOutlined, PoweroffOutlined, RollbackOutlined, UnlockOutlined } from '@ant-design/icons';
 
-import { DeviceToolBarMenu } from 'src/utils/streaming';
+import { DeviceToolBarMenu } from 'src/utils/streaming/streaming';
 import resources from 'src/resources';
 
 interface Props {
