+++ projects/console-web-front/src/utils/streaming/browser-device-inspector.ts
