+++ projects/console-web-front/src/utils/streaming/browser-gamium-service.ts
