+++ projects/console-web-front/src/utils/streaming/streaming.ts
